Deploy a Generative AI solution using a RAG Framework to Google Cloud Challenge Lab L400/requirements.txt
Flask
Jinja2
pytest
pyyaml
google-cloud-aiplatform
firestore
langchain_community
langchain
pypdf
vertexai
